+++ NEWS.md
@@ -1,3 +1,7 @@
 # mdcBHNU 0.0.01
 
-* Initial CRAN submission.
+Updated: 2025-12-30
+
+## New Features
+* Initial setup
+  * added data, filter, and plot functions